pyproject.toml
[tool.poetry]
name = "twitter bot"
version = "0.1.0"
description = ""
authors = ["MauricioFIUBA <mlcabrera@fi.uba.ar>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.31.0"
pylint = "^2.17.4"
python-telegram-bot = "^21.1.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
